// apps/web/src/app/core/models/ApplicationGroupUsersModel.ts
export interface IApplicationGroupUsers {
  applicationGroupUserId: string;
  applicationGroupId: string;
  userId: string;
  applicationId: string;
  status: ApplicationGroupUserStatus;
  createdAt: Date;
  updatedAt: Date;
}